minimalistic reqproc
