--- managed/runSh/run.sh
@@ -1,7 +1,0 @@
-#!/bin/bash -e
-
-main() {
-  node runSh.js
-}
-
-main
